# perf(redis): Multi-threaded and optimized RedisPermissionsRepository