## WHAT

Improves the performance of RedisPermissionsRepository by:

 - Parallelizing critical loops
 - Removing some unnecessary sets
 - Using [`SMEMBERS`](https://redis.io/commands/smembers) instead of [`SSCAN`](https://redis.io/commands/SSCAN) for `scanSet()`

### Benchmarks

Benchmarks were conducted using a custom groovy test which sets up a repository with 10,000 roles, 10,000 users and 100 random roles per-user, then runs a sync including all roles. The numbers below were from running on my m1 macbook pro with 8 CPU cores, 16GB RAM. I ran under both native arm64 and Rosetta, and results were effectively the same.

| change | time |
| --- | --- |
| baseline (master branch 461f759) | 1 minutes, 58.234 seconds |
| getRolesOf parallel only | 1 minutes, 49.630 seconds |
| putAllById parallel only | 1 minutes, 35.675 seconds |
| getAllByRoles parallel only | 58.316 seconds |
| scanSet SMEMBERS only | 28.593 seconds |
| all improvements | 13.638 seconds |

As you can see, the biggest improvement overall is changing `scanSet()` to use `SMEMBERS`, but all of these improvements make a significant impact to the overall runtime of `/roles/sync`. I also tried parallelizing `put()`, but the improvement was negligible relative to these other changes.

I'm not sure of a good place to leave this benchmark in the codebase, so for now I'll just put it here. This should be added to `UserRolesSyncerSpec.groovy`, and run with:

```
./gradlew --info :fiat-roles:test --tests "*should bench*"
```

<details>
<summary>Benchmark Test</summary>
<p>

```groovy
        @Unroll
        def "should bench"() {
            setup:

            Random random = new Random()

            def roleIds = []
            def roles = []
            def users = []
            def userProviderMap = [:]

            for (int i = 0; i < 10000; i++) {
              roles.push(new Role("role" + i).setSource(Role.Source.EXTERNAL))
              roleIds.push("role" + i)
            }

            for (int i = 0; i < 10000; i++) {
              def userRoles = []

              for (int r = 0 ; r < 100; r++) {
                userRoles.push(roles[random.nextInt(10000)])
              }

              def up = new UserPermission()
                      .setId("user" + i)
                      .setAccounts([new Account().setName("account" + i)] as Set)
                      .setRoles(userRoles as Set)

              userProviderMap.put("user" + i, up)

              users.push(up)
            }

            def unrestrictedUser = new UserPermission()
                    .setId(UnrestrictedResourceConfig.UNRESTRICTED_USERNAME)
                    .setAccounts([new Account().setName("unrestrictedAccount")] as Set)

            def abcServiceAcct = new UserPermission().setId("abc").setRoles([roles[0]] as Set)

            userProviderMap.put(UnrestrictedResourceConfig.UNRESTRICTED_USERNAME, unrestrictedUser)
            userProviderMap.put("abc", abcServiceAcct)

            repo.put(unrestrictedUser)
            repo.put(abcServiceAcct)

            for (UserPermission up : users) {
              repo.put(up)
            }

            def permissionsResolver = Mock(PermissionsResolver)
            permissionsResolver.resolve(_ as List) >> {
                userProviderMap
            }
            permissionsResolver.resolveUnrestrictedUser() >> unrestrictedUser

            def lockManager = Mock(LockManager) {
              _ * acquireLock() >> { LockManager.LockOptions lockOptions, Callable onLockAcquiredCallback ->
                onLockAcquiredCallback.call()
              }
            }

            def serviceAccountProvider = Mock(ResourceProvider) {
              getAll() >> [new ServiceAccount().setName("abc").setMemberOf(["role0"])]
            }

            @Subject
            def syncer = new UserRolesSyncer(
                    new DiscoveryStatusListener(true),
                    registry,
                    lockManager,
                    repo,
                    permissionsResolver,
                    serviceAccountProvider,
                    new AlwaysUpHealthIndicator(),
                    1,
                    1,
                    1,
                    1,
                    ""
            )

            when:
            true

            then:
            Date start = new Date()
            syncer.syncAndReturn(roleIds)
            Date stop = new Date()
            TimeDuration td = TimeCategory.minus( stop, start )
            println "Role Sync Duration: " + td
        }

```
</p>
</details>

## WHY

We're still encountering issues with the duration of `/roles/sync` getting longer and longer as we add more users, more applications etc. Ideally, the sync would take less than 20 seconds to keep the sync time less than various timeouts of callers of Fiat.

## TODO

 - [ ] Internal code review
 - [x] Test performance difference in test environment
 - [ ] Test performance difference in production